+++ tests/FluentCertificates.Builder.Tests/SupportedOS.cs
@@ -1,9 +1,0 @@
-﻿namespace FluentCertificates;
-
-public enum SupportedOS
-{
-    FreeBSD = 1,
-    Linux = 2,
-    OSX = 3,
-    Windows = 4,
-}